pyproject.toml
[project]
name = "company"
version = "0.1.0"
description = "Company name extraction from text prompts"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "haystack-ai>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-watch>=4.0.0",
]